Use absolute link for changelog
Hi Kent. I'm currently working on building a changelog parser for Depfu.com and I'm trying my best to autodetect if a project uses releases. Unfortunately, the relative link in this document makes it quite hard to parse the link properly outside of the context that Github has when it renders these pages. My personal view on this is that ideally all documentation text files in a git project should work "standalone" as they might be reused in other contexts. (README's for example are rendered on npmjs.com's package info page, so relative links would fail there as well).

I know this is a bit of a nitpicky change, but I hope you're willing to merge it if only because this would help tools like ours to properly follow the link in an automated fashion without having to laboriously try to mimick the GitHub page context.